cluster/python-scripts/requirements.txt
# Requirements for Python project

# Standard libraries
boto3
json
botocore

# Third-party libraries
pyyaml

# Python built-in libraries
ssl
urllib3
subprocess
os
